DRTII-213 Break out R interface into own object. Tidy tests
